requirements.txt
amqp
asgiref
backports.zoneinfo
billiard
celery
certifi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cron-descriptor
Django
django-celery
django-celery-beat
django-celery-results
django-timezone-field
djangorestframework
idna
kombu
mysqlclient
pika
prompt-toolkit
python-crontab
python-dateutil
pytz
requests
six
sqlparse
typing-extensions
tzdata
urllib3
vine
wcwidth
sqlalchemy
mysql-connector-python